% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pitch_locations.R
\name{plot_pitch_location}
\alias{plot_pitch_location}
\title{\strong{Plot Pitch Locations}}
\usage{
plot_pitch_location(input_df)
}
\arguments{
\item{input_df}{The file you want to use to plot the pitch locations.}
}
\description{
This function allows you to plot all pitch locations from the dataset
}
\examples{
plot_pitch_location()
}
